api/requirements.txt
fastapi==0.104.1
openai==0.27.8
python-dotenv==1.0.0
azure-search-documents==11.3.0
